+++ app/scripts/services/base.serv.js
@@ -25,9 +25,9 @@ angular.module('fdf.services.base', [])
          * @param name
          * @param fn
          */
-        base.evt = function(e, name, fn){
+        base.evt = function(e, fn){
             //发送用户行为
-            base.bahavior(e, name || '');
+            base.bahavior(e);
             //执行回调喊出
             fn && fn.call(null, e);
         };
@@ -85,7 +85,7 @@ angular.module('fdf.services.base', [])
             var e = opts.e;
             var elm = angular.element(opts.elm || e.target);
             // 发送用户行为
-            base.bahavior(e, opts.name);
+            base.bahavior(e, opts);
 
             if(args.length === 1){
                 return;
@@ -216,32 +216,54 @@ angular.module('fdf.services.base', [])
         /**
          * 保存用户行为
          * @param e
-         * @param name
+         * @param opts
          */
-        base.bahavior = function (e, name) {
+        base.bahavior = function (e, opts) {
             //if(!app.SYS.BAHAVIOR.RUN){
             //   return false;
             //}
 
+            // 获得当前用户信息
             var currentUser = base.currentUser();
 
+            // 获得当前DOM信息
             var elm = angular.element(e.target);
-            var moduleElm, moduleName, position;
+            // 获得当前DOM上所绑定的ng scope信息
+            var scope = elm.scope();
 
             // 向上冒泡寻找当前dom所在的module
-            moduleElm = elm.hasClass('fdf-module') ? elm: elm.closest('.fdf-module');
-            moduleName = moduleElm ? moduleElm.attr("fdf-module") : '';
+            var moduleName = base.findModule(scope), duration;
 
             // 获得当前dom的scope，若有$index，则为该dom 所在list 的位置
-            position = elm.scope().$index;
-            position = position != null ? position + 1: null;
+            var position = scope.$index;
+            position = position != null ? position + 1 : null;
 
             // 若evt没传入name，则再到dom中获取
             name = name || elm.attr('fdf-name');
 
+            var lastUrl = app.$rootScope.lastUrl;
+            app.$rootScope.lastUrl = app.$location.absUrl();
+
+            // 若页面跳转，计算页面存在时间
+            if(opts.isChangePage){
+                duration = app.now() - app.$rootScope.startTime;
+                app.$rootScope.startTime = app.now();
+            }
+
             var evtInfo = {
+                // 当前事件类型
                 type: e.type,
+                // 当前事件触发时间
                 timeStamp: e.timeStamp,
+                // 当前事件与上次触发事件的间隔时间
+                intervalTime: app.now() - app.$rootScope.lastTime,
+                // 当前浏览器userAgent 信息
+                userAgent: navigator.userAgent,
+                // 当前用户操作系统信息
+                os: currentUser.os,
+                // 上次打开的页面url
+                lastUrl: lastUrl,
+                // 当前事件信息
                 info: {
                     // 当事件被触发时鼠标指针向对于浏览器页面（或客户区）的坐标
                     clientX: e.clientX,
@@ -255,6 +277,7 @@ angular.module('fdf.services.base', [])
                     // 窗口的左上角在屏幕上的的 x 坐标和 y 坐标 (ie不支持)
                     screenX: e.screenX,
                     screenY: e.ecreenY,
+                    // 当前DOM的信息
                     target: {
                         innerHTML: e.target.innerHTML,
                         tagName: e.target.tagName,
@@ -295,22 +318,69 @@ angular.module('fdf.services.base', [])
             //	页面停留时间（duration）
             // 向后端传递
             var bhinfo = JSON.stringify({
+                // 当前用户ID
                 userId: currentUser.id,
+                // 事件触发事件
                 clickTime: evtInfo.timeStamp,
-                clickPage: app.$rootScope.page,
+                // 事件触发所在页面
+                clickPage: app.$rootScope.channel,
+                // 事件触发的 element 信息
                 clickElement: name,
+                // 事件触发所在模块
                 clickModule: moduleName,
+                // DOM 所在列表中的位置
                 clickPosition: position,
+                // 当前url full path
                 url: app.$location.absUrl(),
-                duration: '',
-                userAgent: '',
-                token: '',
+                // 页面跳出时间
+                duration: duration,
+                // 当前产品名称
+                productName: app.storage('PRODUCT_NAME'),
+                // 当前事件距页面打开时间的间隔时间
+                openToClickTime: app.now() - app.$rootScope.startTime,
+                // 当前用户token
+                token: currentUser.token,
+                // 前端关注的事件信息
                 fe: evtInfo
             });
 
+            app.$log.log(bhinfo);
+
             app.$_Base.bahavior.post({
-                bahavior:bhinfo
+                bahavior: bhinfo
+            }, function(){
+                app.$rootScope.lastTime = app.now();
+            }, function(){
+                app.$rootScope.lastTime = app.now();
             });
+        };
+
+        //通过向上冒泡，寻找当前scope所在最近距离的module信息
+        base.findModule = function(scope){
+            if(!app.isObject(scope)){
+                return null;
+            }
+
+            var key = "module", moduleName, isFind;
+            if(!scope[key]){
+                app.each(scope, function(o){
+                    if(app.type(o, "object")){
+                        if(o[key]){
+                            moduleName = o[key];
+                            isFind = true;
+                            return false;
+                        }
+                    }
+                });
+
+                if(!isFind){
+                    moduleName = base.findModule(scope.$parent);
+                }
+            }else{
+                moduleName = scope[key];
+            }
+
+            return moduleName;
         };
 
         /**